If we have a Splice:
	Source -> Pipe -> Sink

Then if the Pipe is part of a bidirectional connection to Source,
it may send EOS to Sink before the Splice is complete.  Wait until
the Source produces EOS and the Pipe produces EOS to complete the
Splice rather than doing it as soon as the Pipe sends EOS to the
Sink.


git-svn-id: http://wanproxy.org/svn/trunk@640 4068ffdb-0463-0410-8185-8cc71e3bd399